Moved CA functions out into their own module
