+++ mapping/p99/ioc-tests/test_detector_01.py
@@ -1,8 +1,5 @@
-from pkg_resources import require
-require('cothread==2.13')
-from cothread import catools
 import unittest
-
+from ca_common import *
 
 DETECTOR_MODEL = "DC-152Q-C00-FI"
 
@@ -20,11 +17,3 @@ class TestCapture(unittest.TestCase):
     #def test_
     pass
 
-def caget(pv_names):
-    return catools.caget(pv_names)
-
-def caput_async(pv_names, values):
-    catools.caput(pv_names, values)
-
-def caput_sync(pv_names, values):
-    catools.caput(pv_names, values, wait=True)
